refactor: when user has only one game redirect to a game page
