+++ apps/hub/src/routes/_authenticated/_layout/index.tsx
@@ -1,10 +1,9 @@
 import { Intro } from "@/components/intro";
 import { DeepDiveSection } from "@/components/sections/deep-dive-section";
 import { FaqSection } from "@/components/sections/faq-section";
-import { gamesQueryOptions } from "@/domains/game/queries";
-import { ls } from "@/lib/ls";
-import { H1, NarrowPage, Separator, safeAsync } from "@rivet-gg/components";
-import { createFileRoute, redirect } from "@tanstack/react-router";
+import { guardOssNewie } from "@/lib/guards";
+import { H1, NarrowPage, Separator } from "@rivet-gg/components";
+import { createFileRoute } from "@tanstack/react-router";
 import { zodSearchValidator } from "@tanstack/router-zod-adapter";
 import { LayoutGroup } from "framer-motion";
 import { z } from "zod";
@@ -41,24 +40,6 @@ export const Route = createFileRoute("/_authenticated/_layout/")({
       return;
     }
 
-    const lastTeam = ls.get(
-      `rivet-lastteam-${auth.profile?.identity.identityId}`,
-    );
-    if (lastTeam) {
-      throw redirect({
-        to: "/teams/$groupId",
-        params: { groupId: lastTeam },
-      });
-    }
-
-    const [response] = await safeAsync(
-      queryClient.fetchQuery(gamesQueryOptions()),
-    );
-    if (response && response.games.length > 0) {
-      throw redirect({
-        to: "/teams/$groupId",
-        params: { groupId: response.groups[0].groupId },
-      });
-    }
+    await guardOssNewie({ queryClient, auth });
   },
 });
